requirements.txt
pandas
greenlet
sqlalchemy
python-dotenv
beautifulsoup4
urllib3
psycopg2